refactor: split identity-view.tsx and keymodel.ts
Tier 2 item from docs/reports/REPORT_LARGE_FILE_MODULARIZATION_SCAN_2026_07_22.md.

identity-view.tsx: split into identity-account-form.tsx (AccountForm) and
identity-accounts-tab.tsx (AccountsTab/AccountRow/AccountDetail) — two
independent concerns confirmed via grep to have different real consumers
today (AgentIdentityPanel.tsx imports only AccountForm; accounts-manager.tsx
imports both). identity-view.tsx deleted; the two real call sites updated.

keymodel.ts: split into keymodel-debuglog.ts, keymodel-nav.ts,
keymodel-blockcreate.ts, keymodel-dispatch.ts (chord/keymap state, now
exported so registerGlobalKeys can populate it cross-file); keymodel.ts
keeps registerGlobalKeys/countTermBlocks/getAllGlobalKeyBindings. Dropped
3 genuinely-unused imports found along the way (focusManager,
WorkspaceService, deleteLayoutModelForTab).

tsc --noEmit clean; full vitest suite 1904/1908 (one pre-existing unrelated
flaky timeout, confirmed passing in isolation).

<!-- agentmux:agent_id=agentx -->
