+++ frontend/app/store/keymodel-blockcreate.ts
@@ -1,0 +1,64 @@
+// Copyright 2025-2026, AgentMux Corp.
+// SPDX-License-Identifier: Apache-2.0
+
+import { createBlock, createBlockSplitHorizontally, createBlockSplitVertically, getSettingsKeyAtom, WOS } from "@/app/store/global";
+import { getLayoutModelForStaticTab } from "@/layout/index";
+
+export function getDefaultNewBlockDef(): BlockDef {
+    const adnbAtom = getSettingsKeyAtom("app:defaultnewblock");
+    const adnb = adnbAtom() ?? "term";
+    if (adnb == "launcher") {
+        return {
+            meta: {
+                view: "launcher",
+            },
+        };
+    }
+    // "term", blank, anything else, fall back to terminal
+    const termBlockDef: BlockDef = {
+        meta: {
+            view: "term",
+            controller: "shell",
+        },
+    };
+    const layoutModel = getLayoutModelForStaticTab();
+    const focusedNode = layoutModel.focusedNode?.();
+    if (focusedNode != null) {
+        const blockAtom = WOS.getWaveObjectAtom<Block>(WOS.makeORef("block", focusedNode.data?.blockId));
+        const blockData = blockAtom();
+        if (blockData?.meta?.view == "term") {
+            if (blockData?.meta?.["cmd:cwd"] != null) {
+                termBlockDef.meta["cmd:cwd"] = blockData.meta["cmd:cwd"];
+            }
+        }
+        if (blockData?.meta?.connection != null) {
+            termBlockDef.meta.connection = blockData.meta.connection;
+        }
+    }
+    return termBlockDef;
+}
+
+export async function handleCmdN() {
+    const blockDef = getDefaultNewBlockDef();
+    await createBlock(blockDef);
+}
+
+export async function handleSplitHorizontal(position: "before" | "after") {
+    const layoutModel = getLayoutModelForStaticTab();
+    const focusedNode = layoutModel.focusedNode?.();
+    if (focusedNode == null) {
+        return;
+    }
+    const blockDef = getDefaultNewBlockDef();
+    await createBlockSplitHorizontally(blockDef, focusedNode.data.blockId, position);
+}
+
+export async function handleSplitVertical(position: "before" | "after") {
+    const layoutModel = getLayoutModelForStaticTab();
+    const focusedNode = layoutModel.focusedNode?.();
+    if (focusedNode == null) {
+        return;
+    }
+    const blockDef = getDefaultNewBlockDef();
+    await createBlockSplitVertically(blockDef, focusedNode.data.blockId, position);
+}
